requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
openai==1.3.5
aiohttp==3.9.1
requests==2.31.0
